Module Cobol_common

module Basics : sig ... end

module Srcloc : sig ... end

module Copybook : sig ... end

module Diagnostics : sig ... end

module Visitor : sig ... end

module Behaviors : sig ... end

Type tags, useful to name and determine component behaviors

module Tokenizing : sig ... end

exception FatalError of string

val fatal : ('a, Stdlib.Format.formatter, unit, _) Stdlib.format4 -> 'a

module Types : sig ... end

include module type of Types with type 'a res = 'a Types.res and type 'a with_diags = 'a Types.with_diags and type 'a with_loc = 'a Types.with_loc and type lexloc = Types.lexloc and type srcloc = Types.srcloc

type 'a res = 'a Types.res

Transitional (to be removed)

include module type of Diagnostics.TYPES with type 'a with_diags = 'a Types.with_diags

type severity =

| Hint

| Note

| Info

| Warn

| Error

type diagnostic = Diagnostics.TYPES.diagnostic

type diagnostics = Diagnostics.TYPES.diagnostics

type 'a in_result = ('a * diagnostic option, diagnostic option) Stdlib.result

type 'a with_diags = 'a Types.with_diags = {

result : 'a;

diags : diagnostics;

}

include module type of Srcloc.TYPES with type 'a with_loc = 'a Types.with_loc with type lexloc = Types.lexloc with type srcloc = Types.srcloc

type lexloc = Types.lexloc

type srcloc = Types.srcloc

type copylocs = Srcloc.TYPES.copylocs

type 'a with_loc = 'a Types.with_loc = {

payload : 'a;

loc : srcloc;

}

val compare_with_loc : ('a -> 'a -> Ppx_deriving_runtime.int) -> 'a with_loc -> 'a with_loc -> Ppx_deriving_runtime.int

val join_all : ('a, 'e) Stdlib.result list -> ('a list, 'e) Stdlib.result

join_all diags res_l takes every value of a list of ('a,'e)result and return a ('a,'e)listresult. If all the elements in the parameter list are Ok _ then the result is Ok _ otherwise Error _

val join_any : ('a, 'e) Stdlib.result list -> 'a list

join_any diags res_l takes a list of result and makes it a list, the elements of the given list are added to the resulting list only if they are Ok _.

val catch_diagnostics : ?ppf:Stdlib.Format.formatter -> ((module Diagnostics.STATEFUL) -> 'c -> 'a res) -> 'c -> 'a res

catch_diagnostics ?ppf f x enables f to fail abruptly with an exception while operating on an imperative (state-full) diagnostics module.

It first creates a stateful diagnostics module D and calls f D x while intercepting common system exceptions (only Sys_error for now) and fatal diagnostics. An Error is returned in the latter case, or whenever f terminates with an Error.

At last, the diagnostics accumulated in D are printed using ppf (Format.err_formatter by default) in case of a fatal error (fatal is called and FatalError is not caught --- which it shouldn't); in this case FatalError is re-thrown afterwards.

val with_stateful_diagnostics : f:((module Diagnostics.STATEFUL) -> 'b -> 'a) -> 'b -> 'a with_diags

with_stateful_diagnostics ~f x applies f on a fresh stateful diagnostics module D and x, and returns the result combined with all diagnostics accumulated using D.

Any exception that may be raised and escape f is not caught, so it is the responsibility of f to catch it.

val show_diagnostics : ?ppf:Stdlib.Format.formatter -> Diagnostics.Set.t -> unit

val exit : ?status:int -> unit -> _

Exit the program with a status that depends on diagnostics reported in catch_n_show_diagnostics.

val do_unit : ((module Diagnostics.STATEFUL) -> 'a -> unit) -> ?epf:Stdlib.Format.formatter -> 'a -> unit

val do_any : ((module Diagnostics.STATEFUL) -> 'a -> 'b) -> ?epf:Stdlib.Format.formatter -> 'a -> 'b