Module Diagnostics.Set

type t = diagnostics

Persistent set of diagnostics

val pp : t Pretty.printer

val pp_above : level:severity -> t Pretty.printer

val none : t

val one : diagnostic -> t

val maybe : diagnostic option -> t

val two : diagnostic -> diagnostic -> t

val cons : diagnostic -> t -> t

val union : t -> t -> t

val has_errors : t -> bool

val map : (diagnostic -> 'a) -> t -> 'a list

val fold : (diagnostic -> 'a -> 'a) -> t -> 'a -> 'a

val diag : Cobol_common__.Diagnostics_sigs.severity -> ?loc:Srcloc.srcloc -> (_, t) Pretty.func

val hint : ?loc:Srcloc.srcloc -> (_, t) Pretty.func

val note : ?loc:Srcloc.srcloc -> (_, t) Pretty.func

val info : ?loc:Srcloc.srcloc -> (_, t) Pretty.func

val warn : ?loc:Srcloc.srcloc -> (_, t) Pretty.func

val error : ?loc:Srcloc.srcloc -> (_, t) Pretty.func

val blind : diagnostic -> t

type serializable

Sets of diagnostics with values that do not contain any closure or module, and that can thus be marshalled/unmarshalled safely.

The ability to perform delayed formatting is lost when translating from "regular" diagnostics into serializable ones, as the translation applies the formatting as if into a string without any right-margin.

val apply_delayed_formatting : t -> serializable

val of_serializable : serializable -> t