Module Cobol_common.Copybook

type fileloc = [

| `Word of string

| `Alphanum of string

]

type lookup_info = {

libname : string;

libpath : string list;

}

val copybook_extensions : string list

val pp_lookup_error : lookup_info Pretty.printer

val find_lib : libpath:string list -> ?exts:string list -> ?fromfile:string -> ?libname:fileloc -> fileloc -> (string, lookup_info) Stdlib.result

find_lib ~libpath ?exts ?fromfile ?libname txtname attempts to locate a file containing the copybook txtname, which is a file named txtname, possibly appended with an extension from

".CPY"; ".CBL"; ".COB"; ".CBX"; ".cpy";
".cbl"; ".cob"; ".cbx"

(considered in order), unless txtname is given as an alphanumeric literal (e.g, txtname = `Alphanum filname --- in which case no extension is assumed). If libname is not provided, then the file is searched within the directories listed in libpath (considered in order). Otherwise, a single directory libname is considered; if libname is a relative path, it is interpreted relative to the directory that contains fromfile. Lookup is performed in a case-insensitive way. Every directory in the path is considered in order, and all extensions are tried for a given directory of the path in the provided order.